chore(decide): More observability into decide requests
