+++ posthog/api/decide.py
@@ -24,13 +24,13 @@ from posthog.models import Team, User
 from posthog.models.feature_flag import get_all_feature_flags
 from posthog.models.utils import execute_with_timeout
 from posthog.plugins.site import get_decide_site_apps
-from posthog.utils import cors_response, get_ip_address, load_data_from_request
+from posthog.utils import cors_response, get_ip_address, label_for_team_id_to_track, load_data_from_request
 
 
 FLAG_EVALUATION_COUNTER = Counter(
     "flag_evaluation_total",
     "Successful decide requests per team.",
-    labelnames=[LABEL_TEAM_ID, "errors_computing"],
+    labelnames=[LABEL_TEAM_ID, "errors_computing", "has_hash_key_override"],
 )
 
 
@@ -66,29 +66,6 @@ def hostname_in_allowed_url_list(allowed_url_list: Optional[List[str]], hostname
 
 def parse_domain(url: Any) -> Optional[str]:
     return urlparse(url).hostname
-
-
-def label_for_team_id_to_track(team_id: int) -> str:
-    team_id_filter = settings.DECIDE_TRACK_TEAM_IDS
-
-    team_id_as_string = str(team_id)
-
-    if "all" in team_id_filter:
-        return team_id_as_string
-
-    if team_id_as_string in team_id_filter:
-        return team_id_as_string
-
-    team_id_ranges = [team_id_range for team_id_range in team_id_filter if ":" in team_id_range]
-    for range in team_id_ranges:
-        try:
-            start, end = range.split(":")
-            if int(start) <= team_id <= int(end):
-                return team_id_as_string
-        except Exception:
-            pass
-
-    return "unknown"
 
 
 @csrf_exempt
@@ -221,7 +198,11 @@ def get_decide(request: HttpRequest):
 
                 # metrics for feature flags
                 team_id_label = label_for_team_id_to_track(team.pk)
-                FLAG_EVALUATION_COUNTER.labels(team_id=team_id_label, errors_computing=errors).inc()
+                FLAG_EVALUATION_COUNTER.labels(
+                    team_id=team_id_label,
+                    errors_computing=errors,
+                    has_hash_key_override=bool(data.get("$anon_distinct_id")),
+                ).inc()
             else:
                 response["featureFlags"] = {}
