+++ posthog/models/feature_flag/flag_matching.py
@@ -12,6 +12,7 @@ from django.db.models.fields import BooleanField
 from django.db.models import Q
 from django.db.models.query import QuerySet
 from sentry_sdk.api import capture_exception
+from posthog.metrics import LABEL_TEAM_ID
 
 from posthog.models.filters import Filter
 from posthog.models.filters.mixins.utils import cached_property
@@ -24,6 +25,7 @@ from posthog.models.cohort import Cohort
 from posthog.models.utils import execute_with_timeout
 from posthog.queries.base import match_property, properties_to_Q
 from posthog.database_healthcheck import postgres_healthcheck, DATABASE_FOR_FLAG_MATCHING
+from posthog.utils import label_for_team_id_to_track
 
 from .feature_flag import (
     FeatureFlag,
@@ -42,6 +44,12 @@ FLAG_EVALUATION_ERROR_COUNTER = Counter(
     "flag_evaluation_error_total",
     "Failed decide requests with reason.",
     labelnames=["reason"],
+)
+
+FLAG_HASH_KEY_WRITES_COUNTER = Counter(
+    "flag_hash_key_writes_total",
+    "Attempts to write hash key overrides to the database.",
+    labelnames=[LABEL_TEAM_ID, "successful_write"],
 )
 
 
@@ -626,6 +634,8 @@ def get_all_feature_flags(
             writing_hash_key_override = set_feature_flag_hash_key_overrides(
                 team_id, [distinct_id, hash_key_override], hash_key_override
             )
+            team_id_label = label_for_team_id_to_track(team_id)
+            FLAG_HASH_KEY_WRITES_COUNTER.labels(team_id=team_id_label, successful_write=writing_hash_key_override).inc()
         except Exception as e:
             # If the database is in read-only mode, we can't handle experience continuity flags,
             # since the set_feature_flag_hash_key_overrides call will fail.
